New version of gridview.php
Since gridview9.php includes many improvements, I changed the name of
gridview.php to gridviewPrior.php and the name of gridview9.php to
gridview.php.
This version seems to find responses (even before students have
submitted the entire quiz), evaluate the fraction associated with each
response, and create the spreadsheet interface. Fractions are shown as
the background color with red being wrong and light green being correct.
To show these fractions, the value of evaluate must be 1. If showkey=1
the key to the fraction colors is displayed.