Handle pages correctly and include more code.
I changed the code to handle the pages for a question correctly. (The page to be sent depends on the index of the page in the layout  and not on the page in quiz_slots table.) I also moved some code from quiz/atempt.php to this script liveview/pollingattempt.php to make the change to the quiz scripts less.