Modified parocessattempt.php to avoid end of quiz
When the number of answers submitted to a quiz by a student equals the number of questions in a quiz, the quiz program asks the student if the student is ready to finish the quiz. This needed to be eliminated for in-class polling. By modifying quiz/processattempt.php (adding in lines 60-62) this is avoided.