Modify to handle multichoice with checkboxes.
I have modified the code to handle multichoice questions when there are checkboxes and students can select more than one answer. For checkboxes, the step_data has a namechoice(\d) and a value 1 or 0 for select or not. I parse out these, return the asnwer with choces separated by &q& and then explode this into  an array. Each choice counts in the histogram.