Checking for new responses ceases with session timeout
The code in the java script has been changed to find the value of the sessiontimeout value from teh config table and stop looking for new responses after that time has been exceeded. Without this change, some browsers would keep the user logged into the Moodle site as long as the page with this script in it had not been closed.